Fix compiler warnings:
 file included from sketch/Configuration_prusa.h:2:0,
                 from sketch/Configuration.h:43,
                 from sketch/Marlin.h:22,
                 from sketch/Marlin_main.cpp:46:
sketch/Marlin_main.cpp: In function 'void process_commands()':
sketch/variants/1_75mm_MK3-EINSy10a-E3Dv6full.h:106:33: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
 #define NORMAL_MAX_ACCEL_XY     2500ul  // max acceleration in normal mode in mm/s^2
                                 ^
sketch/Marlin_main.cpp:5772:16: note: in expansion of macro 'NORMAL_MAX_ACCEL_XY'
      if (val > NORMAL_MAX_ACCEL_XY)
                ^
sketch/variants/1_75mm_MK3-EINSy10a-E3Dv6full.h:102:34: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
 #define SILENT_MAX_ACCEL_XY      960ul  // max acceleration in silent mode in mm/s^2
                                  ^
sketch/Marlin_main.cpp:5774:23: note: in expansion of macro 'SILENT_MAX_ACCEL_XY'
      if (val_silent > SILENT_MAX_ACCEL_XY)

+++ Firmware/Marlin_main.cpp
@@ -5764,9 +5764,9 @@ Sigma_Exit:
 		{
 			if (code_seen(axis_codes[i]))
 			{
-				int val = code_value();
+				unsigned long val = code_value();
 #ifdef TMC2130
-				int val_silent = val;
+				unsigned long val_silent = val;
 				if ((i == X_AXIS) || (i == Y_AXIS))
 				{
 					if (val > NORMAL_MAX_ACCEL_XY)
